Update Address object
Now it uses aryelgois\utils\Database to handle database
connections (you have to extend this class, because it
is abstract)

And it expects you to have a aryelgois\databases\Address
configured in your server, to provide countries, states
and counties. Other informations are set manually.

This object is useful when you have a registry form and
your clients provide their address. And the <select> for
country, state and county should be filled dynamically.